+++ CHANGELOG.md
@@ -94,6 +94,7 @@ are welcomed to check the E-Mails generated for missed chat messages. @janopae
 - Removed ActivityModel by moving functions to ActivityXhr !1434 #9 @alex.simm 
 - Reimplement footer in vue !1437 @ChrisOelmueller
 - redirected the german refs from storestatusicon.vue and pickuplist.vue to lang.de.yml !1392 #824 @jonathan_b @thesoult
+- Refactored store infos to vue js !1406 @chriswalg
 
 ## Dev/Test/CI stuff
 - Add "linux" tag for finding CI servers !1332 @nicksellen